# internal/controller/handler/v1/carts/merge/gen/type.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CartMergeResponse ログイン時のカート引き継ぎの結果。**引き継ぎで失われた分だけ**を報告します。
// 引き継ぎ後のカートの中身は含めません。取得（GET /v1/carts/me）でいつでも引ける一方、
// ここで失われた分はこの応答でしか得られないためです。
type CartMergeResponse struct {
	// Clamped 数量を合算した結果が 1 明細あたりの上限を超え、上限へ丸めた商品の ID。
	// 引き継ぎは失敗させないため、超過分は捨てて成功として返します。
	//
	//
	// Example: ["b1d4e0f2-3c5a-4b6d-8e7f-1a2b3c4d5e6f"]
	Clamped []openapi_types.UUID `json:"clamped"`

	// Dropped 明細数が上限を超えたため取り込まれなかった商品の ID。
	// 先にカートへ入っていたものを優先して残し、後から来たものを切り捨てます。
	//
	//
	// Example: []
	Dropped []openapi_types.UUID `json:"dropped"`
}

// ErrorResponse エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ErrorResponse struct {
	// Code 機械的に処理可能なエラーコード
	//
	// Example: EXAMPLE_ERROR
	Code string `json:"code"`

	// Message 人間が読めるエラーメッセージ
	//
	// Example: 定義されたエラーメッセージ
	Message string `json:"message"`

	// RequestId リクエストID
	//
	// Example: abcdefghijklmnopqrstuvwxyz
	RequestId string `json:"requestId"`
}

// ErrorResponseWithDetails 詳細識別子付きエラーレスポンススキーマ。 この schema を error レスポンスに宣言したエンドポイントだけが details を返せる（opt-in）。 未宣言のエンドポイントでは errorhandler が details を fail-closed で落とす。
type ErrorResponseWithDetails struct {
	// Code 機械的に処理可能なエラーコード
	//
	// Example: EXAMPLE_ERROR
	Code string `json:"code"`

	// Details エラー詳細を示す公開可能な識別子(任意)。検証エラーでは不正なフィールド名が入る
	//
	// Example: ["firstName","email"]
	Details *[]string `json:"details,omitempty"`

	// Message 人間が読めるエラーメッセージ
	//
	// Example: 定義されたエラーメッセージ
	Message string `json:"message"`

	// RequestId リクエストID
	//
	// Example: abcdefghijklmnopqrstuvwxyz
	RequestId string `json:"requestId"`
}

// CartSessionRequiredParam Example: b3JnLWV4YW1wbGUtc2Vzc2lvbi10b2tlbi1zYW1wbGU
type CartSessionRequiredParam = string

// IdempotencyKeyParam Example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
type IdempotencyKeyParam = string

// BadRequest400 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type BadRequest400 = ErrorResponse

// Conflict409 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type Conflict409 = ErrorResponse

// InternalServerError500 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type InternalServerError500 = ErrorResponse

// MethodNotAllowed405 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type MethodNotAllowed405 = ErrorResponse

// ServiceUnavailable503 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ServiceUnavailable503 = ErrorResponse

// Unauthorized401 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type Unauthorized401 = ErrorResponse

// UnprocessableEntity422 詳細識別子付きエラーレスポンススキーマ。 この schema を error レスポンスに宣言したエンドポイントだけが details を返せる（opt-in）。 未宣言のエンドポイントでは errorhandler が details を fail-closed で落とす。
type UnprocessableEntity422 = ErrorResponseWithDetails

// PostCartsMeMergeParams defines parameters for PostCartsMeMerge.
type PostCartsMeMergeParams struct {
	// XCartSession 引き継ぎ元のゲストカートのセッショントークン。
	// ログイン前に持っていた値をそのまま渡す。
	// 値は 256 ビットを base64url（パディング無し）で表現した 43 文字。
	//
	// 必須である点が CartSessionParam との違い。引き継ぎ元を指定しない引き継ぎは要求として成り立たず、
	// 省略を許すと「引き継ぐものが無い」と「引き継ぎ元を書き忘れた」を区別できなくなる。
	// OpenAPI 3.0 は $ref の兄弟キーを無視するため、共有定義に required を上書きすることはできない。
	XCartSession CartSessionRequiredParam `json:"X-Cart-Session"`

	// IdempotencyKey 冪等キー。非冪等な変更操作（作成系など）で、同一キーによるリトライを重複実行なく
	// 安全にするための識別子。サーバは同一 (認証主体, キー) の再送を初回結果のリプレイとして扱う。
	// このヘッダを宣言する操作は、ハンドラが必ず idempotency.Run 経由で処理する契約（完全性テストで機械検証）。
	IdempotencyKey *IdempotencyKeyParam `json:"Idempotency-Key,omitempty"`
}