# internal/controller/handler/v1/products/count/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetProductsCount 商品検索の一致件数の取得
	// (GET /v1/products/count)
	GetProductsCount(ctx *echo.Context, params GetProductsCountParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetProductsCount converts echo context to params.
func (w *ServerInterfaceWrapper) GetProductsCount(ctx *echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetProductsCountParams
	// ------------- Optional query parameter "categoryId" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "categoryId", ctx.QueryParams(), &params.CategoryId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter categoryId: %s", err))
	}

	// ------------- Optional query parameter "statusId" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "statusId", ctx.QueryParams(), &params.StatusId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter statusId: %s", err))
	}

	// ------------- Optional query parameter "categoryCodes" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "categoryCodes", ctx.QueryParams(), &params.CategoryCodes, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter categoryCodes: %s", err))
	}

	// ------------- Optional query parameter "statusCodes" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "statusCodes", ctx.QueryParams(), &params.StatusCodes, runtime.BindQueryParameterOptions{Type: "array", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter statusCodes: %s", err))
	}

	// ------------- Optional query parameter "keyword" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "keyword", ctx.QueryParams(), &params.Keyword, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter keyword: %s", err))
	}

	// ------------- Optional query parameter "minPrice" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "minPrice", ctx.QueryParams(), &params.MinPrice, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter minPrice: %s", err))
	}

	// ------------- Optional query parameter "maxPrice" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "maxPrice", ctx.QueryParams(), &params.MaxPrice, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter maxPrice: %s", err))
	}

	// ------------- Optional query parameter "minQuantity" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "minQuantity", ctx.QueryParams(), &params.MinQuantity, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter minQuantity: %s", err))
	}

	// ------------- Optional query parameter "maxQuantity" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "maxQuantity", ctx.QueryParams(), &params.MaxQuantity, runtime.BindQueryParameterOptions{Type: "integer", Format: "int32"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter maxQuantity: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProductsCount(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/v1/products/count", wrapper.GetProductsCount, options.OperationMiddlewares["GetProductsCount"]...)

}

type BadRequest400JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type ServiceUnavailable503JSONResponse ErrorResponse

type GetProductsCountRequestObject struct {
	Params GetProductsCountParams
}

type GetProductsCountResponseObject interface {
	VisitGetProductsCountResponse(w http.ResponseWriter) error
}

type GetProductsCount200JSONResponse ProductCountResponse

func (response GetProductsCount200JSONResponse) VisitGetProductsCountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetProductsCount400JSONResponse struct{ BadRequest400JSONResponse }

func (response GetProductsCount400JSONResponse) VisitGetProductsCountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetProductsCount405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response GetProductsCount405JSONResponse) VisitGetProductsCountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type GetProductsCount500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response GetProductsCount500JSONResponse) VisitGetProductsCountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetProductsCount503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response GetProductsCount503JSONResponse) VisitGetProductsCountResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetProductsCount 商品検索の一致件数の取得
	// (GET /v1/products/count)
	GetProductsCount(ctx context.Context, request GetProductsCountRequestObject) (GetProductsCountResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetProductsCount operation middleware
func (sh *strictHandler) GetProductsCount(ctx *echo.Context, params GetProductsCountParams) error {
	var request GetProductsCountRequestObject

	request.Params = params

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetProductsCount(ctx.Request().Context(), request.(GetProductsCountRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetProductsCount")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetProductsCountResponseObject); ok {
		return validResponse.VisitGetProductsCountResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}