# internal/controller/handler/v1/products/ranking/gen/type.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for RankingPeriodParam.
const (
	RankingPeriodParamAll  RankingPeriodParam = "all"
	RankingPeriodParamN30d RankingPeriodParam = "30d"
)

// Valid indicates whether the value is a known member of the RankingPeriodParam enum.
func (e RankingPeriodParam) Valid() bool {
	switch e {
	case RankingPeriodParamAll:
		return true
	case RankingPeriodParamN30d:
		return true
	default:
		return false
	}
}

// Defines values for GetProductsRankingParamsPeriod.
const (
	GetProductsRankingParamsPeriodAll  GetProductsRankingParamsPeriod = "all"
	GetProductsRankingParamsPeriodN30d GetProductsRankingParamsPeriod = "30d"
)

// Valid indicates whether the value is a known member of the GetProductsRankingParamsPeriod enum.
func (e GetProductsRankingParamsPeriod) Valid() bool {
	switch e {
	case GetProductsRankingParamsPeriodAll:
		return true
	case GetProductsRankingParamsPeriodN30d:
		return true
	default:
		return false
	}
}

// ErrorResponse エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ErrorResponse struct {
	// Code 機械的に処理可能なエラーコード
	//
	// Example: EXAMPLE_ERROR
	Code string `json:"code"`

	// Message 人間が読めるエラーメッセージ
	//
	// Example: 定義されたエラーメッセージ
	Message string `json:"message"`

	// RequestId リクエストID
	//
	// Example: abcdefghijklmnopqrstuvwxyz
	RequestId string `json:"requestId"`
}

// ProductRankingItem 売上ランキング 1 商品分の項目です。
type ProductRankingItem struct {
	// Name 商品名
	//
	// Example: ワイヤレスイヤホン
	Name string `json:"name"`

	// Price 価格。サブセント精度を保持する decimal 文字列で表します（例 "19.99" は 19.99 ドル）。 JSON number は IEEE754 double として復元され精度を失うため、文字列で表現します。
	//
	// Example: 19.99
	Price string `json:"price"`

	// ProductId 商品ID
	//
	// Example: b1d4e0f2-3c5a-4b6d-8e7f-1a2b3c4d5e6f
	ProductId openapi_types.UUID `json:"productId"`

	// SoldQuantity 集計期間内の販売数量（キャンセル済みの購入を除く合算値）。
	//
	// Example: 42
	SoldQuantity int64 `json:"soldQuantity"`
}

// ProductRankingResponse 売上ランキングのレスポンススキーマ。販売数量の降順（同数量は商品 ID 昇順）で並びます。
type ProductRankingResponse struct {
	// Rankings 販売数量の降順で並んだ商品ランキングです。
	Rankings []ProductRankingItem `json:"rankings"`
}

// RankingLimitParam Example: 10
type RankingLimitParam = int

// RankingPeriodParam Example: all
type RankingPeriodParam string

// BadRequest400 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type BadRequest400 = ErrorResponse

// InternalServerError500 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type InternalServerError500 = ErrorResponse

// MethodNotAllowed405 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type MethodNotAllowed405 = ErrorResponse

// ServiceUnavailable503 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ServiceUnavailable503 = ErrorResponse

// GetProductsRankingParams defines parameters for GetProductsRanking.
type GetProductsRankingParams struct {
	// Period 集計対象期間。`all`（既定）は全期間、`30d` は注文日時が直近30日以内の購入のみを集計します。
	Period *GetProductsRankingParamsPeriod `form:"period,omitempty" json:"period,omitempty"`

	// Limit 取得する上位件数。売上数量の降順で上位 limit 件を返します。
	Limit *RankingLimitParam `form:"limit,omitempty" json:"limit,omitempty"`
}

// GetProductsRankingParamsPeriod defines parameters for GetProductsRanking.
type GetProductsRankingParamsPeriod string