# internal/controller/handler/v1/purchases/detail/cancel/gen/type.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ErrorResponse struct {
	// Code 機械的に処理可能なエラーコード
	//
	// Example: EXAMPLE_ERROR
	Code string `json:"code"`

	// Message 人間が読めるエラーメッセージ
	//
	// Example: 定義されたエラーメッセージ
	Message string `json:"message"`

	// RequestId リクエストID
	//
	// Example: abcdefghijklmnopqrstuvwxyz
	RequestId string `json:"requestId"`
}

// PurchaseCancelResponse 購入キャンセル後の購入情報レスポンススキーマ。status は ID・名称を事前解決済みで、
// 別途の名称解決 API 呼び出しは不要です。canceledAt にキャンセル日時が入ります。
// 金額（subtotalAmount / taxAmount / shippingFee / totalAmount / 明細 unitPrice）は
// すべて USD セント単位の整数です。
type PurchaseCancelResponse struct {
	// CanceledAt キャンセル日時
	//
	// Example: 2026-07-25T00:00:00Z
	CanceledAt time.Time `json:"canceledAt"`

	// Code 購入コード（利用者へ注文番号として見せる一意の識別子）
	//
	// Example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
	Code string `json:"code"`

	// Details 購入明細の配列
	Details []PurchaseDetailResponse `json:"details"`

	// OrderedAt 注文日時
	//
	// Example: 2026-07-23T00:00:00Z
	OrderedAt time.Time `json:"orderedAt"`

	// ShippingFee 送料。USD セント単位の整数です。
	//
	// Example: 500
	ShippingFee int64 `json:"shippingFee"`

	// Status 購入に紐づくステータス（ID・業務キー・名称）。name は事前解決済みで、別途の名称解決 API 呼び出しは不要です。
	// 表示には name を、分岐にはドメインの業務キーである code を用います。
	Status PurchaseStatusRef `json:"status"`

	// SubtotalAmount 小計。明細（単価×数量）の合計。USD セント単位の整数です。
	//
	// Example: 160000
	SubtotalAmount int64 `json:"subtotalAmount"`

	// TaxAmount 税額（国内消費税）。USD セント単位の整数です。
	//
	// Example: 16000
	TaxAmount int64 `json:"taxAmount"`

	// TotalAmount 合計（小計 + 税額 + 送料）。USD セント単位の整数です。
	//
	// Example: 176500
	TotalAmount int64 `json:"totalAmount"`

	// UserId 購入したユーザーのID
	//
	// Example: 550e8400-e29b-41d4-a716-446655440000
	UserId openapi_types.UUID `json:"userId"`
}

// PurchaseDetailResponse 購入明細のレスポンス。unitPrice は購入時点の価格スナップショットです。
type PurchaseDetailResponse struct {
	// ProductId 商品ID
	//
	// Example: 5266f905-1af4-477a-bc7c-e69729e22a2b
	ProductId openapi_types.UUID `json:"productId"`

	// Quantity 購入数量
	//
	// Example: 2
	Quantity int32 `json:"quantity"`

	// UnitPrice 購入時点の単価スナップショット。サブセント精度を保持する decimal 文字列（USD ドル）で表します（例 "800.00"）。
	// 購入成立後に商品価格が変わっても本値は不変です。
	//
	//
	// Example: 800.00
	UnitPrice string `json:"unitPrice"`
}

// PurchaseStatusRef 購入に紐づくステータス（ID・業務キー・名称）。name は事前解決済みで、別途の名称解決 API 呼び出しは不要です。
// 表示には name を、分岐にはドメインの業務キーである code を用います。
type PurchaseStatusRef struct {
	// Code 購入ステータスの業務キー。値は到達順序を意味しません（完了より支払い済みのほうが大きい）。
	//
	// Example: 7
	Code int64 `json:"code"`

	// Id 購入ステータスID
	//
	// Example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
	Id openapi_types.UUID `json:"id"`

	// Name 購入ステータス名
	//
	// Example: 完了
	Name string `json:"name"`
}

// PurchaseCodeParam Example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
type PurchaseCodeParam = string

// BadRequest400 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type BadRequest400 = ErrorResponse

// Conflict409 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type Conflict409 = ErrorResponse

// InternalServerError500 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type InternalServerError500 = ErrorResponse

// MethodNotAllowed405 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type MethodNotAllowed405 = ErrorResponse

// NotFound404 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type NotFound404 = ErrorResponse

// ServiceUnavailable503 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ServiceUnavailable503 = ErrorResponse

// Unauthorized401 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type Unauthorized401 = ErrorResponse