# openapi/openapi.gen.yaml
openapi: 3.0.3
info:
  title: Go Boilerplate API
  termsOfService: https://github.com/Tomy-ch/go-boilerplate-echo-oapi-sqlc
  contact:
    name: TomyC258460
    url: https://x.com/TomyC258460
  version: 2.2.0
  description: |
    Reference specification for the sample API (users, products, purchases). It exercises every layer
    of the Onion Architecture end to end: cursor and offset pagination, creation and partial update,
    soft delete, idempotency keys, and role-based authorization.

    Handlers (oapi-codegen) and the published reference documentation are both generated from this
    file, so every endpoint change starts here.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
    x-internal: true
  - url: https://api.example.com
    description: Production server (replace with actual URL)
    x-internal: false
  - url: https://staging.api.example.com
    description: Staging server (replace with actual URL)
    x-internal: false
  - url: https://development.api.example.com
    description: Development server (replace with actual URL)
    x-internal: true
security: []
...
paths:
...
  /v1/products/count:
    get:
      operationId: GetProductsCount
      tags:
        - v1/products/count
      summary: 商品検索の一致件数の取得
      description: |
        公開済み商品のうち、指定した検索条件に一致する件数のみを返します。認証不要の公開エンドポイントです。
        categoryCodes / statusCodes / keyword / minPrice / maxPrice / minQuantity / maxQuantity の意味は
        GET /v1/products と同一です。検索条件を指定しない場合は公開済み商品の総数を返します。
        categoryId / statusId は非推奨で、後継の categoryCodes / statusCodes と同時に指定すると 400 を返します。
      security: []
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
        - $ref: '#/components/parameters/StatusIdParam'
        - $ref: '#/components/parameters/CategoryCodesParam'
        - $ref: '#/components/parameters/StatusCodesParam'
        - $ref: '#/components/parameters/KeywordParam'
        - $ref: '#/components/parameters/MinPriceParam'
        - $ref: '#/components/parameters/MaxPriceParam'
        - $ref: '#/components/parameters/MinQuantityParam'
        - $ref: '#/components/parameters/MaxQuantityParam'
      responses:
        '200':
          description: 検索条件に一致する公開商品の件数
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCountResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
components:
  parameters:
...
    KeywordParam:
      name: keyword
      in: query
      description: 全文検索キーワード。商品名・商品説明への部分一致で絞り込みます。
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
        example: example keyword
...
    CategoryIdParam:
      name: categoryId
      in: query
      description: |
        商品カテゴリIDでフィルタします。指定しない場合は全カテゴリを対象とします。
        後継は categoryCodes で、マスタ行を指すのは UUID ではなく code です。categoryCodes と
        同時に指定した場合は 400 を返します。
      required: false
      deprecated: true
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        maxLength: 36
        example: 5dd52d84-78eb-4a52-ba0b-2e11c95c2af2
    StatusIdParam:
      name: statusId
      in: query
      description: |
        商品ステータスIDでフィルタします。指定しない場合は全ステータスを対象とします。
        後継は statusCodes で、マスタ行を指すのは UUID ではなく code です。statusCodes と
        同時に指定した場合は 400 を返します。
      required: false
      deprecated: true
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        maxLength: 36
        example: 093170fb-83a2-4864-a2b3-53236eaf3597
...
    CategoryCodesParam:
      name: categoryCodes
      in: query
      description: |
        商品カテゴリコードでフィルタします（複数指定は同じ名前を繰り返します: `categoryCodes=1&categoryCodes=2`）。
        指定したコードのいずれかに一致する商品を返します。
        指定しない場合は全カテゴリを対象とします。存在しないコードは 0 件として扱い、エラーにはしません。
        コードは商品カテゴリマスタ（GET /v1/products/categories）が返す code で、マスタ行を指す静的な別名です。
        非推奨の categoryId と同時に指定した場合は 400 を返します。
      required: false
      explode: true
      schema:
        type: array
        uniqueItems: true
        maxItems: 32
        items:
          type: integer
          format: int32
          minimum: 1
          maximum: 32767
        example:
          - 1
          - 2
    StatusCodesParam:
      name: statusCodes
      in: query
      description: |
        商品ステータスコードでフィルタします（複数指定は同じ名前を繰り返します: `statusCodes=1&statusCodes=5`）。
        指定したコードのいずれかに一致する商品を返します。
        指定しない場合は全ステータスを対象とします。存在しないコードは 0 件として扱い、エラーにはしません。
        コードは商品ステータスマスタ（GET /v1/products/statuses）が返す code で、マスタ行を指す静的な別名です。
        非推奨の statusId と同時に指定した場合は 400 を返します。
      required: false
      explode: true
      schema:
        type: array
        uniqueItems: true
        maxItems: 32
        items:
          type: integer
          format: int32
          minimum: 1
          maximum: 32767
        example:
          - 1
          - 5
    MinPriceParam:
      name: minPrice
      in: query
      required: false
      description: 最低価格（この値以上）。負でない十進文字列で指定します。
      schema:
        type: string
        maxLength: 40
        pattern: ^\d{1,20}(\.\d{1,18})?$
      example: '10.00'
    MaxPriceParam:
      name: maxPrice
      in: query
      required: false
      description: 最高価格（この値以下）。負でない十進文字列で指定します。
      schema:
        type: string
        maxLength: 40
        pattern: ^\d{1,20}(\.\d{1,18})?$
      example: '100.00'
    MinQuantityParam:
      name: minQuantity
      in: query
      required: false
      description: 最低在庫数（この値以上）。
      schema:
        type: integer
        format: int32
        minimum: 0
      example: 1
    MaxQuantityParam:
      name: maxQuantity
      in: query
      required: false
      description: 最高在庫数（この値以下）。
      schema:
        type: integer
        format: int32
        minimum: 0
      example: 100
...
  securitySchemes:
    BearerAuth:
      type: http
      description: |
        認証にBearerトークンを使用することを示します。
        JWT の検証は RFC8725（JSON Web Token Best Current Practices）に従います。
        署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      description: 認証にBasic認証を使用することを示します。
      scheme: basic
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
...
    ProductCountResponse:
      description: 商品検索の一致件数レスポンス
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          format: int64
          minimum: 0
          description: 検索条件に一致する公開商品の件数。対象がない場合は 0 です。
          example: 42
...
  responses:
    MethodNotAllowed405:
      description: 許可されていないリクエスト方法です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest400:
      description: リクエストが不正です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError500:
      description: サーバーエラーが発生しました。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable503:
      description: サービスが一時的に利用できません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...